Fix crashes when going though a door leading to (0,0) (i.e., exit). - First one was due to Keen 5 specific code being executed in Keen 4. - The second was introduced after redefining CK_object's gfxChunk to be an uint16_t.

+++ src/ck_play.c
@@ -424,7 +424,7 @@ void CK_RunAction(CK_object *obj)
 	{
 		obj->gfxChunk = (obj->xDirection > 0) ? prevAction->chunkRight : prevAction->chunkLeft;
 	}
-	if (obj->gfxChunk == -1)
+	if (obj->gfxChunk == (uint16_t)-1)
 	{
 		obj->gfxChunk = 0;
 	}
